requirements.txt
easyocr
pyperclip
opencv-python
numpy
Pillow
pynput
keyboard